requirements.txt
alembic==1.13.2
bidict==0.23.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pycparser==2.22
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
zipp==3.19.2
